Set default sensible values
Set x sensiblility as twice y, make touchpad joystick more intuitive
as the touchpad is roughly a 2:1 ratio so it does not require using
the full length to reach the emulated joystick full range.

Set sleep_time to half a milisecond, still better than the previous 0ns but
it's really guesswork here.